BUG: Fix `NestedPipeFunction` in graph show wrong datatype (#546)
We cannot resolve the output annotation through the` _NestedPipeFunc.func` property as this is a reference to the `_PipelineAsFunc.call_full_output` method which always annotates the function as `dict[str, Any]`. 

I've added a `output_annotation` property to `NestedPipeFunc` which overwrites the same property of the `PipeFunc` class.
Therefore we can access the `output_annotation` properties of the `PipeFunc`s that constitute the NestedPipeFunc, especially the annotation of the PipeFunc that resembles the output of the `NestedPipeFunc`.

This should resolve  #487.

(There might be an alternative approach by dynamically setting `_PipelineAsFunc.call_full_output` as an attribute with a correct `_PipelineAsFunc.call_full_output.__annotations__` property but I think that the other implementation is a bit cleaner as it seperates the `NestedPipeFunc.output_annotation` logic from the `PipeFunc.output_annotation` logic.)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Bas Nijholt <bas@nijho.lt>